Adding a function for pure morphological analysis
In many cases when the arrays fit directly on GPUs it is easier
and faster to compute the opening using direct morphological operations
using arrayfire's dilate and erode operations. However this places a
limitation on the size of the structuring element as 7x7x7 volumes.
Still it is a good approach because a 1024^3 grid represents a mm
resolution in a meter scale and we can identify features less than 7
mm in this approach. Maybe we recurse on the connected components
after that.